+++ CHANGELOG.md
@@ -5,6 +5,7 @@
   - -- All objectives now utilize the per character objective editor accessed with ctrl+right-click (different than objective templates accessed with the same keybind on the bar anchor)
   - -- Default keybinds have changed for the quick add editboxes
 - Added a mute alert option for individual objectives (character specific)
+- You can now change or remove the template link of an objective on a bar from the objective editor
 - Changed tracker alerts info.trackerObjective to return overall objective by default
 - Added total tracker objective to tooltips in the form:
   - -- trackerName trackerObjectiveCount /  trackerObjective*objectiveObjective (totalTrackerCount / trackerObjective)